state-engine/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
confluent-kafka==2.3.0
redis==5.0.1
pymongo==4.6.0
pydantic==2.5.3
pyyaml==6.0.1
